refactor: move functions to files that make more sense
