+++ src/scikit_hep_repo_review/__init__.py
@@ -9,4 +9,4 @@ from __future__ import annotations
 
 __version__ = "0.5.1"
 
-__all__ = ("__version__",)
+__all__ = ["__version__"]
